Add skeleton DOM implementation
This puts through a skeleton DOM implementation which provide the basis
for running Whiley in the browser.  Using this, other libraries can be
constructed to handle e.g. MVC libraries, etc.

The future of this library is to try and geneate the bindings based on
the living specification and WebIDL.